requirements.txt
Django==4.0.3
djangorestframework==3.13.1
boto3==1.21.17
